requirements.txt
django==5.2
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-filter==25.2
boto3==1.42.4
django-storages==1.14.6
pillow==12.0.0
python-decouple==3.8
psycopg2-binary==2.9.10
coverage==7.12.0
